добавляет д/з 4.16. Больше деталей
