+++ js/main.js
@@ -1,112 +1,3 @@
-/*Функции для создания массива из 25 сгенерированных объектов.
-Каждый объект массива — описание фотографии, опубликованной пользователем.*/
-
-const PICTURES_COUNT = 25;
-const AVATARS_COUNT = 6;
-const COMMENTS_COUNT = 1;
-const LIKES_COUNT = { MIN: 15, MAX: 200 };
-
-const NAMES = [
-  'Зосима',
-  'Лаврентий',
-  'Аглая',
-  'Венедикт',
-  'Иоанн',
-  'Лидия',
-  'Нестор',
-  'Оксана',
-  'Платон',
-  'Радомир',
-  'Серафима',
-  'Трифон',
-  'Устинья',
-  'Феодора',
-  'Харитон',
-  'Целестина',
-  'Чеслава',
-  'Шарлотта',
-  'Элизабета',
-  'Ювеналий',
-  'Ярослава',
-  'Злата',
-  'Измаил',
-  'Казимир',
-  'Лукач',
-];
-
-const COMMENTS = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const DESCRIPTIONS = [
-  'Красивый закат на пляже #природа #закат #пляж',
-  'Портрет кота с красивыми глазами #животные #кот #портрет',
-  'Небоскребы в центре города #город #небоскребы #архитектура',
-  'Цветущее поле под лучами солнца #природа #цветы #лето',
-  'Старинный замок в горах #горы #замок #история',
-  'Водопад в джунглях #природа #водопад #джунгли',
-  'Снежный пейзаж в горах #горы #снег #зима',
-  'Пейзаж с оленями на фоне гор #природа #горы #олени',
-  'Романтический ужин на берегу океана #романтика #океан #пляж',
-  'Абстрактная живопись на стене здания #искусство #граффити #город',
-  'Фотография с высоты птичьего полета над городом #город #вид сверху #птичий полет',
-  'Макрофотография цветов в саду #цветы #макрофотография #сад',
-  'Детский портрет на фоне природы #дети #природа #портрет',
-  'Музыканты на улице вечером #музыка #улица #вечер',
-  'Пейзаж с яркими осенними красками #осень #природа #краски',
-  'Архитектурное чудо в центре города #архитектура #город #здание',
-  'Фотография знаменитого моста в Сан-Франциско #мост #город #Сан-Франциско',
-  'Пейзаж с кораблем на фоне заката #природа #закат #корабль',
-  'Портрет девушки с красивой улыбкой и цветами в руках #портрет #девушка #цветы',
-  'Фотография здания в стиле ар-нуво в центре города #архитектура #город #здание',
-  'Пейзаж с заснеженными горами и лесом вдалеке #горы #снег #лес',
-  'Фотография моста через реку в парке #мост #парк #река',
-  'Пейзаж с яркими цветами на фоне горного озера #природа #озеро #цветы',
-  'Портрет старика с книгой в руках на фоне природы #портрет #старик #природа',
-  'Фотография группы друзей на пляже в летний день #друзья #лето #пляж',
-];
-
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const createIdGenerator = () => {
-  let lastGeneratedId = 0;
-  return () => {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-};
-
-const generatePhotoId = createIdGenerator();
-const generateCommentId = createIdGenerator();
-const generatePhotoUrl = createIdGenerator();
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, AVATARS_COUNT)}.svg`,
-  messages: getRandomArrayElement(COMMENTS),
-  names: getRandomArrayElement(NAMES),
-});
-
-const createPhoto = () => ({
-  id: generatePhotoId(),
-  url: `photos/${generatePhotoUrl()}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(LIKES_COUNT.MIN, LIKES_COUNT.MAX),
-  comments: Array.from({ length: COMMENTS_COUNT }, createComment),
-});
-
-const getPhotos = () => Array.from({ length: PICTURES_COUNT }, createPhoto);
+import {getPhotos} from './data.js';
 
 getPhotos();
